pyproject.toml
[project]
name = "notification-service"
version = "0.1.0"
description = ""
authors = [
    {name = "fodui"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.119.1,<0.120.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "pydantic[email] (>=2.12.3,<3.0.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.17.0,<2.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (==4.3.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pytest (==8.4.2)",
    "pytest-asyncio (>=1.2.0,<2.0.0)",
    "pytest-dotenv (>=0.5.2,<0.6.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
